Coerce to yac_symbol Matrix

The function uses as.yac_symbol() to convert x to an object of class yac_symbol y. If y$is_mat = TRUE, the function returns y. If y$is_vec = TRUE, the function converts y to a column vector, that is, an m by 1 matrix.

as.yac_symbol.mat(x)

as.ysym.mat(x)

Arguments

x
Object.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

as.yac_symbol.mat(1:5)

#> {{1},
#>  {2},
#>  {3},
#>  {4},
#>  {5}} 

(x <- matrix(1:10, ncol = 2))

#>      [,1] [,2]
#> [1,]    1    6
#> [2,]    2    7
#> [3,]    3    8
#> [4,]    4    9
#> [5,]    5   10

as.yac_symbol.mat(x)

#> {{ 1,  6},
#>  { 2,  7},
#>  { 3,  8},
#>  { 4,  9},
#>  { 5, 10}} 

as.ysym.mat(6:10)

#> {{ 6},
#>  { 7},
#>  { 8},
#>  { 9},
#>  {10}} 

(y <- matrix(11:20, ncol = 2))

#>      [,1] [,2]
#> [1,]   11   16
#> [2,]   12   17
#> [3,]   13   18
#> [4,]   14   19
#> [5,]   15   20

as.ysym.mat(y)

#> {{11, 16},
#>  {12, 17},
#>  {13, 18},
#>  {14, 19},
#>  {15, 20}}